added getCitizenshipEligibility and getQualifyingCondition functions in response to issue #8
this is my first stab with javascript and responding to an issue on github. if the functions I added are terrible and unnecessary, I won't be offended! If it's heading in the right direction, I'm happy to look at finishing refactoring getEligibility later unless someone else wants to finish it